Derive: Avoid constructing illegal names
Previously the `DataType` and `Constr` definitions produced by
`deriveDataPrim` would be bound to top-level names formed from the name
of the type whose instance we are deriving. In order to be unique (e.g.
in the case of ByteString) this name needs to include the name of the
defining module. Sadly GHC 7.10 no longer allows names like this. For this
reason we simply bind these names in `where` clauses in `dataTypeOf` and
`toConstr`.